+++ samples/consul-ui/js/service.js
@@ -259,7 +259,7 @@ class ConsulService {
       serviceHtml.appendChild(tagsGenerator(instance.tags));
       serviceHtml.appendChild(serviceMetaGenerator(instance.sMeta));
       serviceHtml.appendChild(connectGenerator(instance))
-      serviceHtml.appendChild(checksStatusGenerator(instance));
+      serviceHtml.appendChild(checksStatusGenerator(instance, instance.name));
       serviceHtml.setAttribute('status', state);
       $("#instances-list").append(serviceHtml);
     }